moved common components from RFC to react package
