--- packages/graphiql-2-rfc-context/src/components/common/Resizer/index.ts
@@ -1,9 +1,0 @@
-/**
- *  Copyright (c) 2021 GraphQL Contributors.
- *
- *  This source code is licensed under the MIT license found in the
- *  LICENSE file in the root directory of this source tree.
- *
- */
-
-export * from './Resizer';
