--- packages/graphiql-2-rfc-context/src/components/common/index.tsx
@@ -1,21 +1,0 @@
-/**
- *  Copyright (c) 2021 GraphQL Contributors.
- *
- *  This source code is licensed under the MIT license found in the
- *  LICENSE file in the root directory of this source tree.
- *
- */
-
-/**
- * Export all the shared components for plugin developers :D
- */
-
-export * from './Logo';
-export * from './Nav';
-
-export * from './List';
-export * from './Toolbar';
-export * from './Toolbar/Tabs';
-export * from './Toolbar/Content';
-export * from './Toolbar/support/WithDividers';
-export * from './Type';
