+++ packages/graphiql-react/src/components/Toolbar/Toolbar.stories.tsx
@@ -11,7 +11,7 @@ import Tabs from './Tabs';
 import React from 'react';
 import Toolbar from './index';
 import Content from './Content';
-import { layout } from '../../themes/decorators';
+import { layout } from '../themes/decorators';
 
 export default { title: 'Toolbar', decorators: [layout] };
