+++ packages/graphiql-react/src/components/themes/default/index.ts
