+++ packages/graphiql-react/src/components/themes/types.ts
